Tooltip Controller
Added a tooltip controller that can show one tooltip at a time, at the mouse position.

To use, grab the instance with ToolTipController.instance, and call SetAndShow(String) to create a tooltip with the given text at the mouse cursor. Will follow the cursor, remember to hide it with Hide()