Updated so towers can be made with multiple resources
If you go to the button image of the tower in the build menu there it will have the MouseTowerCreate script as an component. There you are able to pick the initial costs of the tower. Then in the CellGrid script at the bottom you can change how much the resources increase after each tower placement.